+++ src/main/resources/net/gleske/scmfilter/impl/trait/Messages.properties
@@ -22,5 +22,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
-WildcardSCMFilterTrait.DisplayName=Filter by branch name, tag or pull request (with wildcards)
-RegexSCMFilterTrait.DisplayName=Filter by branch name, tag or pull request (with regular expression)
+WildcardSCMFilterTrait.DisplayName=Filter by branch, tag or pull request (with wildcards)
+RegexSCMFilterBranchTrait.DisplayName=Filter by branch (with regular expression)
+RegexSCMFilterTagTrait.DisplayName=Filter by tag (with regular expression)
+RegexSCMFilterPullRequestDestinationTrait.DisplayName=Filter by pull request destination branch (with regular expression)
+RegexSCMFilterPullRequestOriginTrait_DisplayName=Filter by pull request origin branch (with regular expression)
