+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMFilterPullRequestDestinationTrait/config.jelly
@@ -1,0 +1,6 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+  <f:entry title="${%Pull request destination branch regular expression}" field="prDestinationRegex">
+    <f:textbox default=""/>
+  </f:entry>
+</j:jelly>
