+++ src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMFilterPullRequestOriginTrait/config.jelly
@@ -1,0 +1,6 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+  <f:entry title="${%Pull request origin branch regular expression}" field="prOriginRegex">
+    <f:textbox default=""/>
+  </f:entry>
+</j:jelly>
