--- src/main/resources/net/gleske/scmfilter/impl/trait/RegexSCMFilterTrait/config.jelly
@@ -1,15 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
-  <f:entry title="${%Branch regular expression}" field="branchRegex">
-    <f:textbox default=".*"/>
-  </f:entry>
-  <f:entry title="${%Pull request origin branch regular expression}" field="prOriginRegex">
-    <f:textbox default=""/>
-  </f:entry>
-  <f:entry title="${%Pull request destination branch regular expression}" field="prDestinationRegex">
-    <f:textbox default=""/>
-  </f:entry>
-  <f:entry title="${%Tag regular expression}" field="tagRegex">
-    <f:textbox default="(?!.*)"/>
-  </f:entry>
-</j:jelly>
